Fixed type signature

+++ src/material/color.cc
@@ -7,7 +7,7 @@ namespace material {
     return (1.0-t)*from  + t*to;
   }
 
-  std::tuple<int, int, int> rgbColor(const Color &c) {
+  std::tuple<long, long, long> rgbColor(const Color &c) {
     return {long(255.99*c.x), long(255.99*c.y), long(255.99*c.z)};
   };
 }
